Move Abbreviation class in to languages common
